+++ public/js/filetree.js
@@ -2,32 +2,63 @@
  * d3 filetree for diskover-web
  */
 
-$('#submit').click(function () {
-	console.log('changing paths');
-	var newdir = $('#path').val();
-	var filter = $_GET('filter');
-	var maxdepth = $_GET('maxdepth');
-	location.href = "/sunburst.php?path=" + newdir + "&filter=" + filter + '&maxdepth=' + maxdepth;
-	return false;
+$(document).ready(function () {
+
+	$('#submit').click(function () {
+		console.log('changing paths');
+		var newdir = $('#path').val();
+		var filter = parseInt($_GET('filter'));
+		location.href = "/filetree.php?path=" + newdir + "&filter=" + filter;
+		return false;
+	});
+
 });
 
-function showHidden(data) {
-	// display warning if no files or some other error
-	if (data == null || data.error) {
-		console.warn("nothing found in Elasticsearch");
-		document.getElementById('error').style.display = 'block';
-		return false;
-	}
+function showHidden(root) {
 	// data is loaded so let's show hidden elements on the page
 	// update path field
-	document.getElementById('path').value = jsondata.name;
+	document.getElementById('path').value = root.name;
 	// show path input
 	document.getElementById('path-container').style.display = 'inline-block';
 	// show filetree buttons
 	document.getElementById('buttons-container').style.display = 'inline-block';
-	// show sunburst div
-	document.getElementById('sunburst-container').style.display = 'block';
-	return true;
+	// show chart div
+	document.getElementById('chart-container').style.display = 'block';
+}
+
+function getChildJSON(d) {
+	// get json data from Elasticsearch using php data grabber
+	console.log("getting children from Elasticsearch");
+
+	var target = document.getElementById('tree-container');
+	// trigger loader
+	var spinner = new Spinner(opts).spin(target);
+
+	// url for d3.json
+	var url = '/d3_data.php?path=' + d.name + '&filter=' + filter;
+
+	// load json data and trigger callback
+	d3.json(url, function (error, data) {
+
+		// display error if data has error message
+		if ((data && data.error) || error) {
+			spinner.stop();
+			console.warn("nothing found in Elasticsearch: " + error);
+			document.getElementById('error').style.display = 'block';
+			return false;
+		}
+
+		if (data.children.length > 0) {
+			// update children in root
+			d._children = [];
+			d._children = data.children;
+		}
+
+		// stop spin.js loader
+		spinner.stop();
+
+	});
+
 }
 
 function getJSON() {
@@ -35,59 +66,46 @@ function getJSON() {
 	console.time('loadtime')
 
 	// check if json data stored in session storage
-	data = JSON.parse(sessionStorage.getItem("diskover-filetree"));
-	root = JSON.parse(sessionStorage.getItem("diskover-filetree-root"));
+	root = JSON.parse(sessionStorage.getItem("diskover-filetree"));
 
 	// get data from Elasticsearh if no json in session storage
-	if (!data && !root) {
-		getESJsonData(nodata = true);
-		return true;
-	} else if (!data) {
+	if (!root) {
 		getESJsonData();
 		return true;
 	}
 
 	// get new json data if filter cookies are different than current url params
-	if ($_GET('maxdepth') != '' && $_GET('filter') != '') {
-		if ($_GET('maxdepth') != getCookie('maxdepth') || $_GET('filter') != getCookie('filter')) {
-			console.log("removing json data on local storage because filters changed");
+	if ($_GET('filter') != '') {
+		if ($_GET('filter') != getCookie('filter')) {
+			console.log("removing json data on local storage because filter changed");
 			sessionStorage.removeItem("diskover-filetree");
-			sessionStorage.removeItem("diskover-filetree-root");
-			getESJsonData(nodata = true);
+			getESJsonData();
 			return true;
 		}
 	}
 
-	// if path is root, get root json data
-	if (root && path == root.name) {
-		console.log("path is root so using json root data");
-		data = root;
-		sessionStorage.setItem('diskover-filetree', JSON.stringify(data));
-		loadVisualizations();
-		return true;
-	}
-
 	// get new json data from ES if path changed
-	if (data.name != path) {
+	if (root.name != path) {
 		console.log("removing json data on local storage because path changed");
 		sessionStorage.removeItem("diskover-filetree");
 		getESJsonData();
 		return true;
-	} else if (data.name == path) {
+	} else if (root.name == path) {
 		// json data on local storage is same as path so lets show the visuals
+		console.log("json data in storage same as path, load visuals");
 		loadVisualizations();
 		return true;
 	}
 
-	function getESJsonData(nodata = false) {
+	function getESJsonData() {
 		// get json data from Elasticsearch using php data grabber
-		console.log("no json data in storage, grabbing from Elasticsearch");
+		console.log("no json data in session storage, grabbing from Elasticsearch");
 
 		// trigger loader
 		var spinner = new Spinner(opts).spin(target);
 
-		// load json data and trigger callback
-		d3.json(chartConfig.data_url, function (error, json) {
+		// load json data from Elasticsearch
+		d3.json(chartConfig.data_url, function (error, data) {
 
 			// display error if data has error message
 			if ((data && data.error) || error) {
@@ -97,22 +115,18 @@ function getJSON() {
 				return false;
 			}
 
-			// store json from Elasticsearch into data
-			data = json;
+			root = data;
 
-			if (nodata == true && path == "/") {
-				console.log("storing root path json data in storage");
-				// store in session Storage
-				sessionStorage.setItem('diskover-filetree-root', JSON.stringify(data));
-			}
-
-			console.log("storing path json data in storage");
+			console.log("storing json data in session storage");
 			// store in session Storage
 			sessionStorage.setItem('diskover-filetree', JSON.stringify(data));
 
 			// stop spin.js loader
 			spinner.stop();
+			
+			console.timeEnd('loadtime');
 
+			// load d3 visuals
 			loadVisualizations();
 
 		});
@@ -120,48 +134,26 @@ function getJSON() {
 
 	function loadVisualizations() {
 
-		// return error if we haven't found any json data
-		if (!data || data.error) {
-			console.warn("path not found in json data");
-			document.getElementById('error').style.display = 'block';
-			return false;
-		}
-
-		// jsondata for sunburst
-		jsondata = JSON.parse(JSON.stringify(data));
-		// make copy
-		jsondata0 = JSON.parse(JSON.stringify(data));
-
-		console.log("JSON:");
-		console.log(jsondata);
-
-		showHidden(data);
+		// show hidden elements on page
+		showHidden(root);
 
 		// store cookies
-		setCookie('path', $('#path').val()); // decodeURIComponent($_GET('path'))
+		setCookie('path', $('#path').val());
 		($_GET('filter')) ? setCookie('filter', $_GET('filter')): setCookie('filter', 1048576);
-		($_GET('maxdepth')) ? setCookie('maxdepth', $_GET('maxdepth')): setCookie('maxdepth', 3);
 
 		// update file tree link
 		changeFileTreeLink();
 
 		// load file tree
-		updateTree(data, data, firstrun = true);
-		// load Sunburst
-		onJson(null, jsondata);
+		updateTree(root, root);
+		
+		// load pie chart
+		changePie(root);
 	}
-
-	console.timeEnd('loadtime');
 
 }
 
-function updateTree(data, parent, firstrun) {
-
-	// setup tree if first time
-	if (firstrun) setupTree(data);
-
-	var nodes = tree.nodes(data),
-		duration = 250;
+function updateTree(data, parent) {
 
 	function toggleChildren(d) {
 		if (d.children) {
@@ -173,50 +165,8 @@ function updateTree(data, parent, firstrun) {
 		}
 	}
 
-	function getSunChild(d) {
-		// directory changed so we need to get child from sunburst json
-		var sunchild;
-
-		// if at root return
-		if (d.name == jsondata.name) return jsondata;
-
-		// loop through all the children to find child (d)
-		jsondata.children.forEach(getChild);
-		jsondata._children.forEach(getChild);
-
-		function getChild(child) {
-			if (child.name == d.name && child.parent.name == d.parent.name) {
-				sunchild = child;
-				return sunchild;
-			}
-			// recurse if child has children
-			if (child.children) {
-				child.children.forEach(getChild);
-			}
-		}
-		return sunchild;
-	}
-
-	function updateSunburst(d) {
-		loc0 = d.name;
-		if (d.children || d._children) {
-			// directory changed, update sunburst
-			console.log("update sunburst");
-			// at root
-			if (d.depth == 0) {
-				console.log('AT ROOT');
-				var root = getSunChild(d);
-				console.log(root);
-				zoom(root, root);
-			} else {
-				var sunchild = getSunChild(d);
-				console.log(sunchild);
-				zoom(sunchild, sunchild);
-			}
-			return true;
-		}
-		return false;
-	}
+	var nodes = tree.nodes(data),
+		duration = 250;
 
 	var nodeEls = ul.selectAll("li.node").data(nodes, function (d) {
 		d.id = d.id || ++id;
@@ -229,11 +179,14 @@ function updateTree(data, parent, firstrun) {
 		.style("opacity", 0)
 		.style("height", tree.nodeHeight() + "px")
 		.on("click", function (d) {
-			toggleChildren(d);
-			updateTree(data, d, firstrun = false);
-			loc = d.name;
-			// update sunburst
-			if (loc0 != loc && d.depth <= maxdepth - 2) updateSunburst(d);
+			if (d.count > 0 && !d.children && !d._children) {
+				// check if there are any children in Elasticsearch
+				getChildJSON(d);
+			} else if (d.children || d._children) {
+				toggleChildren(d);
+				updateTree(data, d);
+				changePie(d);
+			}
 		})
 		.on("mouseover", function (d) {
 			d3.select(this).classed("selected", true);
@@ -256,15 +209,16 @@ function updateTree(data, parent, firstrun) {
 	//add text
 	entered.append("span").attr("class", "filename")
 		.html(function (d) {
-			return d.name;
+			return d.depth == 0 ? d.name : d.name.split('/').pop();
 		});
 	//add filesize
 	entered.append("span").attr("class", function (d) {
-			if (d.size > 10737418240) {
+			var percent = (d.size / (d.parent || root).size * 100).toFixed(0);
+			if (percent >= 90) {
 				var fileclass = "filesize-red";
-			} else if (d.size > 5368709120 && d.size < 10737418240) {
+			} else if (percent >= 75) {
 				var fileclass = "filesize-orange";
-			} else if (d.size > 1073741824 && d.size < 5368709120) {
+			} else if (percent >= 50) {
 				var fileclass = "filesize-yellow";
 			} else {
 				var fileclass = "filesize-gray";
@@ -274,10 +228,16 @@ function updateTree(data, parent, firstrun) {
 		.html(function (d) {
 			return format(d.size);
 		});
+	// add percent
+	entered.append("span").attr("class", "percent")
+		.style("width", function (d) {
+			var percent = (d.size / (d.parent || root).size * 100).toFixed(0);
+			return percent + "%";
+		});
 	// add file count
 	entered.append("span").attr("class", "filecount")
 		.html(function (d) {
-			return d.count;
+			return d.count > 0 ? "(" + d.count + ")" : "";
 		});
 	//update caret arrow direction
 	nodeEls.select("span.downarrow").attr("class", function (d) {
@@ -297,42 +257,15 @@ function updateTree(data, parent, firstrun) {
 	nodeEls.exit().remove();
 }
 
-function setupTree(data) {
-
-	function collapse(d) {
-		if (d.children) {
-			d._children = d.children;
-			d._children.forEach(collapse);
-			d.children = null;
-		}
-	}
-
-	function expandSingle(d) {
-		if (d._children) {
-			if (d.depth == 0) {
-				d.children = d._children;
-				d._children = null;
-			}
-		}
-	}
-	data.children.forEach(collapse);
-	data.children.forEach(expandSingle);
-}
-
-var data,
-	jsondata,
-	jsondata0,
-	id = 0,
-	loc0,
-	loc;
+var root,
+	id = 0;
 
 var tree = d3.layout.treelist()
 	.childIndent(15)
 	.nodeHeight(18);
 var ul = d3.select("#tree-container").append("ul").classed("treelist", "true");
 
-var filter = $_GET('filter');
-var maxdepth = $_GET('maxdepth');
+var filter = parseInt($_GET('filter'));
 var path = decodeURIComponent($_GET('path'));
 // remove any trailing slash
 if (path != '/') {
@@ -341,12 +274,15 @@ if (path != '/') {
 
 console.log("PATH:" + path);
 console.log("FILTER:" + filter);
-console.log("MAXDEPTH:" + maxdepth);
+
+// add filter and maxdepth to statustext
+var status_filter = ($_GET('filter')) ? 'filter:' + format($_GET('filter')) : 'filter:1 MB';
+document.getElementById('statusfilters').append(status_filter);
 
 // config references
 var chartConfig = {
 	target: 'mainwindow',
-	data_url: '/d3_data.php?path=' + path + '&filter=' + filter + '&maxdepth=' + maxdepth
+	data_url: '/d3_data.php?path=' + path + '&filter=' + filter
 };
 
 // loader settings
@@ -364,11 +300,6 @@ var opts = {
 // loader settings
 var target = document.getElementById(chartConfig.target);
 
-// add filter and maxdepth to statustext
-var status_filter = ($_GET('filter')) ? 'filter:' + format($_GET('filter')) : 'filter:1M';
-document.getElementById('statusfilters').append(status_filter);
-var status_maxdepth = ($_GET('maxdepth')) ? ' maxdepth:' + $_GET('maxdepth') : ' maxdepth:3';
-document.getElementById('statusfilters').append(status_maxdepth);
 
-// get json data and load file tree and sunburst
+// get json data
 getJSON();